requirements.txt
requests
pycryptodome
python-dotenv
pywin32
